+++ src/SaleMode.cs
@@ -1,0 +1,8 @@
+﻿namespace WeekdayBonuses;
+
+public enum SaleMode
+{
+	Random,
+	Highest,
+	Custom
+}