+++ CHANGELOG.md
@@ -1,2 +1,11 @@
+### v1.1.0
+- Add Double Monster Monday
+- Add Double Trap Tuesday
+- Move Small Facility to Thursday
+- Add Black Friday
+- Add Nightmare Saturday
+- Add config for Small Facility to apply a multiplier
+- Add config to have multiple events on one day
+
 ### v1.0.0
 - Initial release